// src/app/shared/models/chart.model.ts
export interface IChartWithGrid {
    chartData: any;
    gridData: any;
}

export interface ChartOptions {
    series: any;
    chart: any;
    xaxis: any;
    yaxis: any;
    title: any;
    theme: any;
    stroke: any;
    dataLabels: any;
    fill: any;
    tooltip: any;
    plotOptions: any;
    legend: any;
    labels: any;
    responsive: any;
    colors: any;
    markers: any;
    subtitle: any;
    icon: string;
    noData: any;
    grid: any;
}